+++ src/dearevenue.jl
@@ -8,6 +8,7 @@ struct RevenueDEAModel <: AbstractRevenueDEAModel
     m::Int64
     s::Int64
     rts::Symbol
+    disposX::Symbol
     eff::Vector
     lambda::SparseMatrixCSC{Float64, Int64}
     techeff::Vector
@@ -22,7 +23,7 @@ inputs `X`, outputs `Y` and price of outputs `P`.
 
 # Optional Arguments
 - `rts=:VRS`: chooses variable returns to scale. For constant returns to scale choose `:CRS`.
-- `disposal=:Strong`: chooses strong disposal of inputs. For weak disposal choose `:Weak`.
+- `dispos=:Strong`: chooses strong disposability of inputs. For weak disposability choose `:Weak`.
 
 # Examples
 ```jldoctest
@@ -47,7 +48,7 @@ Orientation = Output; Returns to Scale = VRS
 ──────────────────────────────────
 ```
 """
-function dearevenue(X::Matrix, Y::Matrix, P::Matrix; rts::Symbol = :VRS, disposal::Symbol = :Strong)::RevenueDEAModel
+function dearevenue(X::Matrix, Y::Matrix, P::Matrix; rts::Symbol = :VRS, dispos::Symbol = :Strong)::RevenueDEAModel
     # Check parameters
     nx, m = size(X)
     ny, s = size(Y)
@@ -64,8 +65,8 @@ function dearevenue(X::Matrix, Y::Matrix, P::Matrix; rts::Symbol = :VRS, disposa
         error("number of output prices and outputs is different")
     end
 
-    if disposal != :Strong && disposal != :Weak
-        error("Invalued disposal $disposal. Disposal should be :Strong or :Weak")
+    if dispos != :Strong && dispos != :Weak
+        error("Invalued disposability $dispos. Disposability should be :Strong or :Weak")
     end
 
     # Compute efficiency for each DMU
@@ -87,9 +88,9 @@ function dearevenue(X::Matrix, Y::Matrix, P::Matrix; rts::Symbol = :VRS, disposa
 
         @objective(deamodel, Max, sum(p0[j] .* Yeff[j] for j in 1:s))
 
-        if disposal == :Strong
+        if dispos == :Strong
             @constraint(deamodel, [j in 1:m], sum(X[t,j] * lambda[t] for t in 1:n) <= x0[j])
-        elseif disposal == :Weak
+        elseif dispos == :Weak
             @constraint(deamodel, [j in 1:m], sum(X[t,j] * lambda[t] for t in 1:n) == x0[j])
         end
 
@@ -119,28 +120,28 @@ function dearevenue(X::Matrix, Y::Matrix, P::Matrix; rts::Symbol = :VRS, disposa
 
     # Revenue, technical and allocative efficiency
     refficiency  = vec( sum(P .* Y, dims = 2) ./ sum(P .* Yefficient, dims = 2) )
-    techefficiency = 1 ./ efficiency(dea(X, Y, orient = :Output, rts = rts, slack = false, disposalX = disposal))
+    techefficiency = 1 ./ efficiency(dea(X, Y, orient = :Output, rts = rts, slack = false, disposX = dispos))
     allocefficiency = refficiency ./ techefficiency
-    return RevenueDEAModel(n, m, s, rts, refficiency, rlambdaeff, techefficiency, allocefficiency)
+    return RevenueDEAModel(n, m, s, rts, dispos, refficiency, rlambdaeff, techefficiency, allocefficiency)
 
 end
 
-function dearevenue(X::Vector, Y::Matrix, P::Matrix, rts::Symbol = :VRS, disposal::Symbol = :Strong)::RevenueDEAModel
+function dearevenue(X::Vector, Y::Matrix, P::Matrix, rts::Symbol = :VRS, dispos::Symbol = :Strong)::RevenueDEAModel
     X = X[:,:]
-    return dearevenue(X, Y, P, rts = rts, disposal = disposal)
+    return dearevenue(X, Y, P, rts = rts, dispos = dispos)
 end
 
-function dearevenue(X::Matrix, Y::Vector, P::Vector; rts::Symbol = :VRS, disposal::Symbol = :Strong)::RevenueDEAModel
+function dearevenue(X::Matrix, Y::Vector, P::Vector; rts::Symbol = :VRS, dispos::Symbol = :Strong)::RevenueDEAModel
     Y = Y[:,:]
     P = P[:,:]
-    return dearevenue(X, Y, P, rts = rts, disposal = disposal)
+    return dearevenue(X, Y, P, rts = rts, dispos = dispos)
 end
 
-function dearevenue(X::Vector, Y::Vector, P::Vector; rts::Symbol = :VRS, disposal::Symbol = :Strong)::RevenueDEAModel
+function dearevenue(X::Vector, Y::Vector, P::Vector; rts::Symbol = :VRS, dispos::Symbol = :Strong)::RevenueDEAModel
     X = X[:,:]
     Y = Y[:,:]
     P = P[:,:]
-    return dearevenue(X, Y, P, rts = rts, disposal = disposal)
+    return dearevenue(X, Y, P, rts = rts, dispos = dispos)
 end
 
 function Base.show(io::IO, x::RevenueDEAModel)
@@ -149,6 +150,8 @@ function Base.show(io::IO, x::RevenueDEAModel)
     n = nobs(x)
     m = ninputs(x)
     s = noutputs(x)
+    disposX = x.disposX
+
     eff = efficiency(x)
     techeff = efficiency(x, :Technical)
     alloceff = efficiency(x, :Allocative)
@@ -162,9 +165,9 @@ function Base.show(io::IO, x::RevenueDEAModel)
         print(io, "Orientation = Output")
         print(io, "; Returns to Scale = ", string(x.rts))
         print(io, "\n")
+        if disposX == :Weak println("Weak disposability of inputs") end
+
         show(io, CoefTable(hcat(eff, techeff, alloceff), ["Revenue", "Technical", "Allocative"], ["$i" for i in 1:n]))
-
-    else
-
     end
+
 end