+++ test/dea.jl
@@ -240,11 +240,11 @@
     @test_throws ErrorException dea([1; 2; 3], [4; 5; 6], orient = :Error) # Invalid orientation
     @test_throws ErrorException dea([1; 2; 3], [4; 5; 6], rts = :Error) # Invalid returns to scale
 
-    @test_throws ErrorException dea([1; 2 ; 3], [4 ; 5; 6], disposalX = :Error)  # Invalid inputs disposal
-    @test_throws ErrorException dea([1; 2 ; 3], [4 ; 5; 6], disposalY = :Error)  # Invalid outputs disposal
+    @test_throws ErrorException dea([1; 2 ; 3], [4 ; 5; 6], disposX = :Error)  # Invalid inputs disposability
+    @test_throws ErrorException dea([1; 2 ; 3], [4 ; 5; 6], disposY = :Error)  # Invalid outputs disposability
 
     # ------------------
-    # Weak Diposal Tests
+    # Weak Disposability Tests
     # ------------------
 
     X = [1; 2; 3; 2; 4]
@@ -255,7 +255,7 @@
     @test slacks(deaioStrong, :X) ≈ [0; 0; 0; 0; 0] atol=1e-15
     @test slacks(deaioStrong, :Y) ≈ [0; 0; 0; 1; 0] atol=1e-15
 
-    deaioWeak = dea(X, Y, orient = :Input, rts = :VRS, disposalY = :Weak)
+    deaioWeak = dea(X, Y, orient = :Input, rts = :VRS, disposY = :Weak)
     @test efficiency(deaioWeak ) ≈ [1.0; 1.0; 1.0; 1.0; 0.5]
     @test slacks(deaioWeak, :X) ≈ [0; 0; 0; 0; 0] atol=1e-15
     @test slacks(deaioWeak, :Y) ≈ [0; 0; 0; 0; 0] atol=1e-15
@@ -265,7 +265,7 @@
     @test slacks(deaooStrong, :X) ≈ [0; 0; 0; 0; 1] atol=1e-15
     @test slacks(deaooStrong, :Y) ≈ [0; 0; 0; 0; 0] atol=1e-15
 
-    deaooWeak = dea(X, Y, orient = :Output, rts = :VRS, disposalX = :Weak)
+    deaooWeak = dea(X, Y, orient = :Output, rts = :VRS, disposX = :Weak)
     @test efficiency(deaooWeak ) ≈ [1.0; 1.0; 1.0; 3.0; 1.0]
     @test slacks(deaooWeak, :X) ≈ [0; 0; 0; 0; 0] atol=1e-14
     @test slacks(deaooWeak, :Y) ≈ [0; 0; 0; 0; 0] atol=1e-14